Update 1_README.md
I think this provides a basic outline of what to do. It will be important to update the SOP as a step by step work through of classification (and TransectMeasure).
There is not much mentioned of TM. First mentioned half way through. Was this on purpose? I understand this classification can be used in lots of programs that would do a similar job but I think it may want to emphasized that we do it in TM up front.